CHI-469 - [Case View] - Loading indicator
