+++ plugin-hrm-form/src/components/case/Case.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/prop-types */
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Template, ITask } from '@twilio/flex-ui';
 import { connect, ConnectedProps } from 'react-redux';
 import { CircularProgress } from '@material-ui/core';
@@ -7,6 +7,7 @@ import AddIcon from '@material-ui/icons/Add';
 import CancelIcon from '@material-ui/icons/Cancel';
 import Dialog from '@material-ui/core/Dialog';
 import DialogContent from '@material-ui/core/DialogContent';
+import { format } from 'date-fns';
 
 import { StandaloneITask } from '../StandaloneSearch';
 import {
@@ -19,7 +20,8 @@ import {
 } from '../../states';
 import { getConfig } from '../../HrmFormPlugin';
 import { saveToHrm, connectToCase, transformCategories } from '../../services/ContactService';
-import { cancelCase, updateCase } from '../../services/CaseService';
+import { cancelCase, updateCase, getActivities } from '../../services/CaseService';
+import { isConnectedCaseActivity, getDateFromNotSavedContact, sortActivities } from './caseHelpers';
 import { Box, BottomButtonBar, StyledNextStepButton } from '../../styles/HrmStyles';
 import { CaseContainer, CenteredContainer } from '../../styles/case';
 import CaseDetails from './CaseDetails';
@@ -82,8 +84,44 @@ const Case: React.FC<Props> = props => {
   const [anchorEl, setAnchorEl] = useState(null);
   const [isMenuOpen, setMenuOpen] = useState(false);
   const [mockedMessage, setMockedMessage] = useState(null);
-  const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(true);
+  const [timeline, setTimeline] = useState([]);
   const { route, subroute } = props.routing;
+
+  useEffect(() => {
+    /**
+     * Gets the activities timeline from current caseId
+     * If the case is just being created, adds the case's description as a new activity.
+     */
+    const getTimeline = async () => {
+      const { task, form, connectedCaseState } = props;
+
+      const activities = await getActivities(connectedCaseState.connectedCase.id, setLoading);
+      let timelineActivities = sortActivities(activities);
+
+      const isCreatingCase = !timelineActivities.some(isConnectedCaseActivity);
+
+      if (isCreatingCase) {
+        if (isStandaloneITask(task)) return;
+        const date = getDateFromNotSavedContact(task, form);
+        const { workerSid } = getConfig();
+        const connectCaseActivity = {
+          date: format(date, 'yyyy-MM-dd HH:mm:ss'),
+          createdAt: new Date().toISOString(),
+          type: task.channelType,
+          text: form.caseInformation.callSummary,
+          twilioWorkerId: workerSid,
+          channel: task.channelType === 'default' ? form.contactlessTask.channel : task.channelType,
+        };
+
+        timelineActivities = sortActivities([...timelineActivities, connectCaseActivity]);
+      }
+      console.log(JSON.stringify(timelineActivities));
+      setTimeline(timelineActivities);
+    };
+
+    getTimeline();
+  }, [props, setLoading]);
 
   const toggleCaseMenu = e => {
     e.persist();
@@ -308,7 +346,7 @@ const Case: React.FC<Props> = props => {
               />
             </Box>
             <Box marginLeft="25px" marginTop="25px">
-              <Timeline caseObj={connectedCase} task={task} form={form} status={status} />
+              <Timeline timeline={timeline} caseObj={connectedCase} task={task} form={form} status={status} />
             </Box>
             <Box marginLeft="25px" marginTop="25px">
               <Households
